requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
daytona-sdk>=0.125.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.0.0
httpx>=0.25.0
